Fixed OMP calls, print, and graphs
The omp call in "solveC" and in "setInitialCond" was incorrectly typed
as "!#omp" instead of the proper "!$omp". This is now fixed.

There is now a final print once tEnd has been reached. Effectivly
printing the final timestep.

The 3D shell scripts where edited so that the gnuplot creates a square
aspect ratio for its plots.