Fixed System variables
Changed the variables/parameters of f(C) and g(C). Now they are correct.
As a result of this lots of subroutine calls had to be changed.
Essentially what changed was:
f(C) is now = \mu * \frac{C}{k + C} - \nu
and
g(C) is now = y \frac{C}{k + C}
so these, \mu, \nu, and y have new parameter values, as shown in
parameter.txt, which was changed accordingly.

Also added an exit condition in the iterations between solutions, so
that if there is no convergence after 100 iterations, the program exits.
Hopefully this never actually actives, but it's nice for a peace of
mind.