Massive parameter/f-g-Select cleanup
With the previous update fixing the parameters of the system, this one
changed the parameter file so that there is no longer dimensional
variables. Before the dimensional value would be given and the
non-dimensionalization would occur within the program (use the dimen
values to create a non-dimensional value). This was not ideal, so now
there are just the 4 non-dimensional values, to which a specific value
is given. We do not need to worry about trying to obtain a dimensional
value from these results (at least not yet).

Along with the system parameters, there were lots of program or
numerical method specific variables that were removed. They are:

length - removed since it was only used in non-dimensionalisation,
instead of being used to change the region of simulations. If needed,
this can be re-implemented when that feature is needed (changing
region).

lambda, xLen, yLen, yDel, (xDel) - These were removed (changed), since
it was inteded to be for varying grid shapes, but since the intended
behaviour is to have uniform grids, these were no longer needed.
Specifically, lambda, xLen were never used; yLen was alwasy equal to 1;
yDel and xDel were the same, so now just xDel is used.

m0, c0 - Removed since they were just for non-dimensionlisation

gam - Removed since it was only used in a deleted fSelect

(fSelect, gSelect) - The unused selects were deleted. For gSelect, only
gSelect 1 was used the rest are delted. fSelect, only monod and contois
kinetic was used, so the rest (1 - 11 before) were deleted. Now the
existing fSelect are renamed to 1 - 4.

Also, the variable names were changed to better reflect their actural
naming convention. [previous name -> Current Name]
        y       -> gama
        esoln   -> eSoln
        k       -> kappa

A few other minor changes were made. Mostly just fixing the subroutine
headers to account for the changes.

This was tested and gives results. Now just need to find suitable
parameter values.