Fixed 2D initial condition and cleaned-up gama
The 2D initial condition wouldn't run, the way it was written with i and
j was a little archaic, so it was changed to the way of the 3D initial
condition. Now instead of only computing the values for which I known to
be positive and non-zero, we calculate every grid point with the initial
conditions and then force any negative value to be zero. More
calculations are obviously required, but at this point it is an
insignificant increase to computation time.

The cleaned-up gama refers to the stdout of the parameter values, gama
was mis-aligned, and to the parameter.txt file, which now defaults gama
to a values that allows the biomass to reach a density of ~0.95, so that
the diffusion has a chance to do something.