Trav wave detection subroutines
This is the start of a method to detect if the current solution is a
travelling wave solution based on the solution from 10 timesteps
previous (10 is subject to change).

The idea here is to have a new output file "travcheck.dat" that has two
columns: t and trav exist. Where t is the time at that output and trav
exist is a boolean value (0,1) for if a trav wave exists at that point.
Then, based on this output file. If every timestep after a certain time,
t_{trav}, is a travelling wave, then we know that a travelling wave
solution exists.

The main changes so far is the addition of a 'checkTrav' parameter,
which changes the outputs of the program to include the
"travecheck.dat". Along with this is the new subroutine 'checkTravWave'.
This is the current plroblem (doesn't work), but the idea is to take two
different solutions of M and check if each datat point is c +- eTrav
away from each other. The assumption is that c is the first difference
between points, but the trouble is that the data points are not frequent
enough to get the grids that I need.

The current plan is to try and increase pSize to see if that helps...
Maybe have a tolerance that is affected by pSize or at least the value
of row....