Added random innoculation point IC
Added MinitialCond = 4 and 5, which are random innoculation points over
the whole domaina and on the y=0 side, respectivly.

The randomness of the points is controled by the init_random_seed
subroutine, which seeds to the system_clock, so it should be pretty
random between runs. This new subroutine is in the new file,
"init_random_seed.f90".

This change was accompanied by the addition of a "innocPts" parameter,
which depicts the number of innoculation points to be used.

There is one major change that should be done to this initialcondition,
the total initial biomass should be monitored. This would mean tracking
how much biomass there is after each innoculation point either by
adding the exact amount for that innoculation point or by calculation
the whole region after each addition, the latter being easier but more
costly. The last innoculation point would have the height/depth changed
so that it would equate the final amount of biomass needed to reach a
constant amount between simulations. The arbitrary amount chosen should
be (pi/2*depth^2 * height), which is the total biomass for
MinitialCond==3. THis can be changed later.

Also, with the above addition, there should be some randomness between
inoculation points, so that they have mildly different shapes (like an
epsilon (10%) of difference for depth and height).

And there is one problem with this, innoculation point will overwrite
other innoculation points if they overlap. This can be resolved by
adding the values together (i.e. M(j) += blah blah or by checking if the
randomly seleceted x and y will result in a overlap....